refactor: Player와 Dealer는 Participant를 상속하도록 변경
- 코드 재사용성 증가
- Player와 Dealer와 다른 메서드 하나를 오버라이드해서 다른 행동을 하도록 지정가능
