+++ src/main/java/blackjack/domain/Dealer.java
@@ -1,0 +1,14 @@
+package blackjack.domain;
+
+public class Dealer extends Participant {
+
+    public Dealer(String name) {
+        super(name);
+    }
+
+    @Override
+    public boolean canDraw() {
+        return calculateCardsValue() < 17;
+    }
+
+}
